add exponential function for opencl benchmark

+++ opencl/calc_utils.c
@@ -48,7 +48,7 @@ double diffArray(const int datasize, const double *data1, const double *data2)
   return (diff_sum);
 }
 
-double timeFunc(void (*func)(const int datasize, const double *data1, const double *data2, double *result),
+double timeFunc(void (*func)(const unsigned long datasize, const double *data1, const double *data2, double *result),
 		const int loop,
 		const int datasize,
 		const double *data1,
