+++ opencl/calc_utils.h
@@ -6,7 +6,7 @@
 double getTime();
 void setRandomData(const int size, double *data);
 double diffArray(const int datasize, const double *data1, const double *data2);
-double timeFunc(void (*func)(const int datasize, const double *data1, const double *data2, double *result),
+double timeFunc(void (*func)(const unsigned long datasize, const double *data1, const double *data2, double *result),
 		const int loop,
 		const int datasize,
 		const double *data1,
